+++ tests/0300_views/0311_consumable.py
@@ -18,12 +18,17 @@ def test_index_empty(client, db):
     assert len(dom.find(".consumable-index .consumables .item")) == 0
 
 
-def test_index_filled(client, db, initial_catalog):  # noqa: F811
+def test_index_filled(client, db, initial_catalog,  # noqa: F811
+                      django_assert_num_queries):
     """
     Consumable index should list all available consumables.
     """
     url = reverse("atoum:consumable-index")
-    response = client.get(url, follow=True)
+
+    # Only a single queryset to list objects
+    with django_assert_num_queries(1):
+        response = client.get(url, follow=True)
+
     assert response.redirect_chain == []
     assert response.status_code == 200
 
@@ -32,13 +37,19 @@ def test_index_filled(client, db, initial_catalog):  # noqa: F811
     assert titles == ["Food", "Hygiene", "Other consumable", "Pets"]
 
 
-def test_detail_filled(client, db, initial_catalog):  # noqa: F811
+def test_detail_filled(client, db, initial_catalog,  # noqa: F811
+                       django_assert_num_queries):
     """
     Consumable detail should list its related assortments.
     """
     foods = initial_catalog.consumables["foods"]
     url = reverse("atoum:consumable-detail", kwargs={"slug": foods.slug})
-    response = client.get(url, follow=True)
+
+    # A queryset for the main object, another one to list its related objects and
+    # another one for pagination
+    with django_assert_num_queries(3):
+        response = client.get(url, follow=True)
+
     assert response.redirect_chain == []
     assert response.status_code == 200